LLM Project/requirements.txt
langchain==0.0.339
python-dotenv==1.0.0
streamlit==1.29.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=4.25.1